# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local-first activity and memory tracking API for OpenClaw.",
        "title": "Clawtivity API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/activity": {
            "get": {
                "description": "List activity entries with optional filters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "List activities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by project_tag",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by model",
                        "name": "model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by created_at date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.ActivityFeed"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new activity entry from OpenClaw hook payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Create activity",
                "parameters": [
                    {
                        "description": "Activity data",
                        "name": "activity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.ActivityFeed"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.ActivityFeed"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.APIError"
                        }
                    }
                }
            }
        },
        "/api/activity/summary": {
            "get": {
                "description": "Get aggregated activity stats with optional filters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Get activity summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by project_tag",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by model",
                        "name": "model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by created_at date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.ActivitySummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.APIError"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns current service/database health details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "database.ActivityFeed": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "channel": {
                    "type": "string"
                },
                "cost_estimate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_ms": {
                    "type": "integer"
                },
                "external_ref": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "project_tag": {
                    "type": "string"
                },
                "reasoning": {
                    "type": "boolean"
                },
                "session_key": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "thinking": {
                    "type": "string"
                },
                "tokens_in": {
                    "type": "integer"
                },
                "tokens_out": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "database.ActivitySummary": {
            "type": "object",
            "properties": {
                "by_status": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "cost_total": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "duration_ms_total": {
                    "type": "integer"
                },
                "tokens_in_total": {
                    "type": "integer"
                },
                "tokens_out_total": {
                    "type": "integer"
                }
            }
        },
        "server.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}